backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
spacy==3.7.2
transformers==4.35.2
torch==2.2.0
python-magic==0.4.27
Pillow==10.1.0
openpyxl==3.1.2
sqlalchemy==2.0.23
python-multipart==0.0.6
textstat==0.7.3
langdetect==1.0.9
nltk==3.8.1
regex==2023.10.3
pdfplumber==0.9.0
python-dotenv==1.0.0
openai==1.3.0 